add process_pkt_logic

+++ tests/netstar/playground.cc
@@ -661,12 +661,12 @@ public:
             //To do list:
             //schedule the task, following is the strategy offload all to GPU
             sort(_flows.begin(),_flows.end(),CompLess);
-            uint64_t partition=get_partition();
+            int partition=get_partition();
             gpu_pkts=(char*)malloc(partition*_flows[partition-1]->packets.size()*sizeof(char*));
             gpu_states=(char*)malloc(partition*sizeof(char*));
-            for(uint64_t i=0; i<partition; i++){
+            for(int i=0; i<partition; i++){
                 gpu_states[i]=reinterpret_cast<char*>(&(_flows[i]->_fs));
-                for(uint64_t j=0;j<_flows[i]->packets.size();j++){
+                for(int j=0;j<(int)_flows[i]->packets.size();j++){
                     gpu_pkts[i][j]=reinterpret_cast<char*>(_flows[i]->packets[j].get_header<net::eth_hdr>(0));
                 }
             }
